backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pymupdf
google-genai
pinecone
psycopg2-binary
python-dotenv
